pyproject.toml
[project]
name = "ozon-mcp"
version = "0.6.0"
description = "Ozon Seller API & Performance API knowledge-rich MCP server for AI agents"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "PCDCK", url = "https://github.com/PCDCK" }]
requires-python = ">=3.12"
keywords = ["mcp", "ozon", "marketplace", "ai-agents", "claude", "openapi", "seller-api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
]

dependencies = [
    "mcp>=1.2.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "rank-bm25>=0.2.2",
    "structlog>=24.4.0",
    "networkx>=3.3",
    "pyyaml>=6.0.2",
    "snowballstemmer>=2.2.0",
    "httpx>=0.27.2",
    "tenacity>=9.0.0",
    "aiolimiter>=1.1.0",
    "jsonschema>=4.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-httpx>=0.32.0",
    "respx>=0.21.1",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "types-pyyaml",
    "types-jsonschema",
]

[project.scripts]
ozon-mcp = "ozon_mcp.__main__:main"

[project.urls]
Homepage = "https://github.com/PCDCK/ozon-mcp"
Repository = "https://github.com/PCDCK/ozon-mcp"
Issues = "https://github.com/PCDCK/ozon-mcp/issues"
Changelog = "https://github.com/PCDCK/ozon-mcp/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ozon_mcp"]

[tool.hatch.build.targets.wheel.force-include]
"src/ozon_mcp/data/seller_swagger.json" = "ozon_mcp/data/seller_swagger.json"
"src/ozon_mcp/data/perf_swagger.json" = "ozon_mcp/data/perf_swagger.json"
"src/ozon_mcp/data/swagger_meta.json" = "ozon_mcp/data/swagger_meta.json"
"src/ozon_mcp/knowledge/workflows.yaml" = "ozon_mcp/knowledge/workflows.yaml"
"src/ozon_mcp/knowledge/rate_limits.yaml" = "ozon_mcp/knowledge/rate_limits.yaml"
"src/ozon_mcp/knowledge/error_codes.yaml" = "ozon_mcp/knowledge/error_codes.yaml"
"src/ozon_mcp/knowledge/quirks.yaml" = "ozon_mcp/knowledge/quirks.yaml"
"src/ozon_mcp/knowledge/examples.yaml" = "ozon_mcp/knowledge/examples.yaml"
"src/ozon_mcp/knowledge/safety_overrides.yaml" = "ozon_mcp/knowledge/safety_overrides.yaml"
"src/ozon_mcp/knowledge/subscription_overrides.yaml" = "ozon_mcp/knowledge/subscription_overrides.yaml"
"src/ozon_mcp/knowledge/deprecated_methods.yaml" = "ozon_mcp/knowledge/deprecated_methods.yaml"
"src/ozon_mcp/knowledge/pagination_patterns.yaml" = "ozon_mcp/knowledge/pagination_patterns.yaml"
"src/ozon_mcp/knowledge/descriptions_overrides.yaml" = "ozon_mcp/knowledge/descriptions_overrides.yaml"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
# Russian-language test fixtures and docstrings legitimately contain Cyrillic
# letters that visually overlap with Latin (с / s). Suppress the ambiguous-
# unicode rule for tests where this is intentional.
"tests/**" = ["RUF001"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests/unit", "tests/golden", "tests/integration"]
addopts = "--tb=short"
markers = [
    "live: tests requiring real Ozon API credentials (deselect with '-m \"not live\"')",
]

[tool.coverage.run]
source = ["src/ozon_mcp"]
omit = [
    "*/tests/*",
    "*/__main__.py",
]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 70
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
